fix script startup deadlock, add script params to create storage and index, fix fs store
To build an index, use this:
moin index-build -s -i
moin index-build --storage-create --index-create

Usually moin would expect an existing index and storage when starting up
and fail if it is not there. This is a problem for moin index-* commands,
which are used to build an index.

By giving -s, you force it to create the storage where specified by the config.
By giving -i, you force it to create the index where specified by the config.

fs storage used os.mkdir. use os.makedirs to support multiple levels of dirs
getting create. catch exception if the dir is already there.

--HG--
branch : storage-ng