New routing middleware based on namespaces, details see below.
Dispatch to different backends based on a namespace, so e.g.
SomePage -> use content/default backend
userprofiles:JoeDoe -> use userprofile backend

If you do not give a NAMESPACE item in the metadata, it will parse the NAME item
and create the NAMESPACE item (and modify the NAME item to only have the name
without namespace).

Uses 3 mappings:
* namespace_mapping: namespace prefix -> backend_name
* backend_mapping: backend_name -> Backend
* acl mapping: maps namespace:itemname prefixes to sets of acls

namespace_mapping and backend_mapping are kept separate so we can map different
namespaces into same backend (this is possible as NAMESPACE key is stored in
meta data). We also need to have a backend name anyway for serialization.

ACLs might not use it correctly right now (one has to give namespace:itemname)
to the acl check.

--HG--
branch : namespaces