avoid crash on password change submit if we have no enc_password stored
can happen e.g. when using GivenAuth to just login in (and autocreate)
some specific user. that user won't have a enc_password then.

for the password change form, we require the user to give the old password
and the new one. if we have no enc_password stored, that should just fail
(and it does fail now, but not crash).

--HG--
branch : namespaces