indexing middleware: use correct logger, prepare code for fixing name(s) issues
I changed the workarounds for the current issues related to name(s) values so
that they log warnings when they are used. So we do not hide the issues, but
are aware that some code still needs fixing. I also streamlined the name(s)
properties so they work in the same way at both places.

If one see some Workaround warning and wants to work on it, one can just
uncomment the "raise ValueError" to see where exactly it is (mis)used.

revision_meta[NAME] or whoosh_doc[NAME] should always be a list of unicode (not: a single
unicode value).

the .names property should return just that list in the end (when all places
are fixed).

the .name property maybe should return either some preferred name (unicode) or
None, if there is no name (empty NAME list).

If we have fixed all places, we can later remove the workarounds and use assertions
or raise ValueErrors/TypeErrors.

--HG--
branch : namespaces